Code-repository/Cross-modal retrieval beyond text-image/Text-motion cross-modal retrieval/TMR++/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.1.1
cmake==3.25.0
colorlog==6.7.0
einops==0.6.1
filelock==3.9.0
frozenlist==1.4.0
fsspec==2023.9.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
idna==3.4
Jinja2==3.1.2
lightning-utilities==0.9.0
lit==15.0.7
MarkupSafe==2.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.0
numpy==1.24.1
omegaconf==2.3.0
orjson==3.9.7
packaging==23.1
Pillow==9.3.0
pytorch-lightning==2.0.9
PyYAML==6.0.1
requests==2.31.0
scipy==1.11.2
sympy==1.11.1
torchmetrics==1.1.2
transformers==4.41.2
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
yarl==1.9.2
